Backend/requirements.txt
Pillow
my-voice-analysis
pycountry
django
djangorestframework
django-cors-headers